# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/machines": {
      "get": {
        "operationId": "MachinesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List all machines",
        "tags": [
          "Machines"
        ]
      },
      "post": {
        "operationId": "MachinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMachineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create a machine",
        "tags": [
          "Machines"
        ]
      }
    },
    "/machines/{id}": {
      "get": {
        "operationId": "MachinesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get machine by ID",
        "tags": [
          "Machines"
        ]
      }
    },
    "/machines/{id}/meter": {
      "patch": {
        "operationId": "MachinesController_updateMeter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Update machine meter reading",
        "tags": [
          "Machines"
        ]
      }
    },
...
    "/operators": {
      "get": {
        "operationId": "OperatorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List operators",
        "tags": [
          "Operators"
        ]
      },
      "post": {
        "operationId": "OperatorsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create an operator",
        "tags": [
          "Operators"
        ]
      }
    },
    "/operators/{id}": {
      "get": {
        "operationId": "OperatorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get operator by ID",
        "tags": [
          "Operators"
        ]
      },
      "put": {
        "operationId": "OperatorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Update an operator",
        "tags": [
          "Operators"
        ]
      }
    },
...
    "/maintenance/machines/{machineId}/tasks": {
      "get": {
        "operationId": "MaintenanceController_getTasks",
        "parameters": [
          {
            "name": "machineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get maintenance tasks for a machine",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/maintenance/machines/{machineId}/status": {
      "get": {
        "operationId": "MaintenanceController_getStatus",
        "parameters": [
          {
            "name": "machineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get maintenance status for a machine",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/maintenance/machines/{machineId}/visits": {
      "get": {
        "operationId": "MaintenanceController_getVisits",
        "parameters": [
          {
            "name": "machineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get maintenance visits for a machine",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/maintenance/visits": {
      "post": {
        "operationId": "MaintenanceController_createVisit",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create a maintenance visit",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/expenses": {
      "get": {
        "operationId": "ExpensesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List expenses",
        "tags": [
          "Expenses"
        ]
      },
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create an expense",
        "tags": [
          "Expenses"
        ]
      }
    },
    "/expenses/categories": {
      "get": {
        "operationId": "ExpensesController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List expense categories",
        "tags": [
          "Expenses"
        ]
      }
    },
    "/expenses/{id}/corrections": {
      "post": {
        "operationId": "ExpensesController_correct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Correct an expense (creates new version, e.g. void with reason)",
        "tags": [
          "Expenses"
        ]
      }
    },
...
    "/client-money/events": {
      "get": {
        "operationId": "ClientMoneyController_getEvents",
        "parameters": [
          {
            "name": "client_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List client money events",
        "tags": [
          "Client Money"
        ]
      },
      "post": {
        "operationId": "ClientMoneyController_createEvent",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create a client money event (receipt/advance/credit note)",
        "tags": [
          "Client Money"
        ]
      }
    },
    "/client-money/receivables": {
      "get": {
        "operationId": "ClientMoneyController_getReceivables",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get client receivables",
        "tags": [
          "Client Money"
        ]
      }
    },
    "/client-money/unused-advances": {
      "get": {
        "operationId": "ClientMoneyController_getUnusedAdvances",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get unused advances",
        "tags": [
          "Client Money"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet OS API",
    "description": "Multi-tenant SaaS for heavy-equipment operators",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateMachineDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "EXC-001"
          },
          "type": {
            "type": "string",
            "example": "excavator"
          },
          "make": {
            "type": "string",
            "example": "Caterpillar"
          },
          "model": {
            "type": "string",
            "example": "320F"
          },
          "year": {
            "type": "number",
            "example": 2020
          },
          "chassis_no": {
            "type": "string"
          },
          "primary_meter_type": {
            "type": "string",
            "example": "hours"
          },
          "meter_unit_label": {
            "type": "string",
            "example": "hours"
          },
          "client_uuid": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "type",
          "primary_meter_type",
          "meter_unit_label",
          "client_uuid"
        ]
      },
...
    }
  }
}